+++ packages/good-master-data/build.js
@@ -1,0 +1,7 @@
+// import cp from 'child_process';
+// import chalk from 'chalk';
+
+// const { execSync } = cp;
+// const envLog = (str) => console.log(chalk.yellow.bold(str));
+// envLog(`${execSync('git --version')}`.replace('\n', ''));
+// envLog(`${execSync('mv dist/packages/**/*.d.ts dist')}`.replace('\n', ''));