+++ packages/button/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to this project will be documented in this file. See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.9](https://github.com/harry118/winshare-ui/compare/@winshare/button@1.1.8...@winshare/button@1.1.9) (2022-09-05)
+
+**Note:** Version bump only for package @winshare/button
+
+## [1.1.8](https://github.com/harry118/winshare-ui/compare/@winshare/button@1.1.7...@winshare/button@1.1.8) (2022-09-01)
+
+**Note:** Version bump only for package @winshare/button
+
 ## [1.1.6](https://github.com/harry118/winshare-ui/compare/@winshare/button@1.1.5...@winshare/button@1.1.6) (2022-08-16)
 
 **Note:** Version bump only for package @winshare/button